requirements.txt
# Core packages
python-dotenv==1.0.1

# LangChain-related packages
langchain-chroma==0.1.4
langchain-graph-retriever==0.8.0
langchain-google-genai==0.0.4
langchain==0.1.0

# ML/DL Framework
torch==2.0.0
torchvision==0.15.0
scikit-learn==1.3.2
numpy==1.24.3
matplotlib==3.8.0
tqdm==4.66.1

# Dataset utilities
datasets==2.14.0
huggingface-hub==0.17.0
transformers==4.30.0

# Graph Database
neo4j==5.14.0

# Supporting dependencies
requests==2.31.0
urllib3==2.0.4
certifi==2024.12.14
charset-normalizer==3.3.1
idna==3.4
